Refactor code for search bar
