+++ .gitignore
@@ -22,4 +22,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-.secret.js
+
+
+# Local Netlify folder
+.netlify
