+++ src/index.css
@@ -6,7 +6,6 @@ body {
   height: 100vh;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  /* text-transform: uppercase; */
 }
 
 code {
